Improve javascript code and styles
