+++ src/public/js/components/productTopInfoSection.js
@@ -4,52 +4,20 @@ import lgFullscreen from 'lightgallery/plugins/fullscreen/lg-fullscreen.min.js';
 import lgThumbnail from 'lightgallery/plugins/thumbnail/lg-thumbnail.min.js';
 import lgVideo from 'lightgallery/plugins/video/lg-video.min.js';
 import lgZoom from 'lightgallery/plugins/zoom/lg-zoom.min.js';
-import Plyr from 'plyr';
-import Swiper from 'swiper/bundle';
+import { updateCounters, updateMainContent } from './contentUpdate';
+import { initializeSwiper } from './swiper';
 
-const SWIPER_SELECTOR = '.swiper';
 const SWIPER_SLIDE_SELECTOR = '.swiper-slide';
 const SWIPER_SELECTED_SLIDE_SELECTOR = '.swiper-slide--selected';
 const MOBILE_MEDIA_QUERY = '(min-width: 64rem)';
 
-const SWIPER_CONFIG = {
-  speed: 1000,
-  loopFillGroupWithBlank: false,
-  direction: 'horizontal',
-  loop: true,
-  slidesPerView: 1,
-  slidesPerGroup: 1,
-  loopFillGroupBlank: false,
-  spaceBetween: 7,
-  breakpoints: {
-    1200: {
-      loop: false,
-      slidesPerView: 6,
-      slidesPerGroup: 6,
-    },
-    1024: {
-      loop: false,
-      slidesPerView: 5,
-      slidesPerGroup: 5,
-    },
-  },
-  pagination: {
-    el: '.swiper-pagination',
-    clickable: true,
-  },
-  navigation: {
-    nextEl: '.swiper-button-next',
-    prevEl: '.swiper-button-prev',
-  },
-};
-
 let preventSlideClick = false;
-let activePlyr = null;
 let currentIndex = 0;
 const slides = document.querySelectorAll(SWIPER_SLIDE_SELECTOR);
 const swiperInstance = initializeSwiper();
-initializeSlideClickListeners(slides);
 const lightGalleryInstance = initializeLightGallery();
+
+initializeSlideClickListeners(slides);
 initializeMediaQueryListener();
 setActiveSlide(document.querySelector(SWIPER_SELECTED_SLIDE_SELECTOR));
 
@@ -64,11 +32,6 @@ new Drift(document.querySelector('.carousel__main-img'), {
 document.querySelector('.carousel__active-element').addEventListener('click', () => {
   lightGalleryInstance.openGallery(currentIndex);
 });
-
-function initializeSwiper() {
-  const swiper = new Swiper(SWIPER_SELECTOR, SWIPER_CONFIG);
-  return swiper;
-}
 
 function initializeSlideClickListeners(slides) {
   slides.forEach((slide) => {
@@ -140,55 +103,6 @@ function updateCarouselUI(selectedSlide) {
     index: selectedSlide.getAttribute('data-index'),
   };
   currentIndex = srcProps.index;
-  updateCounters(srcProps.index);
+  updateCounters(srcProps.index, slides);
   updateMainContent(srcProps);
-}
-
-function updateCounters(index) {
-  const counters = document.querySelectorAll('.carousel__counter');
-  counters.forEach((counter) => {
-    counter.textContent = `${formatNumber(+index + 1)} / ${formatNumber(slides.length)}`;
-  });
-}
-
-function updateMainContent(srcProps) {
-  const mainImg = document.querySelector('.carousel__main-img');
-  const mainVideoContainer = document.querySelector('.carousel__main-video-container');
-  const mainVideo = mainVideoContainer.querySelector('.carousel__main-video');
-
-  if (srcProps.video) {
-    updateVideoContent(mainVideo, mainVideoContainer, mainImg, srcProps.video);
-  } else {
-    updateImageContent(mainImg, mainVideoContainer, srcProps.src);
-  }
-}
-
-function updateVideoContent(mainVideo, mainVideoContainer, mainImg, videoInfo) {
-  if (!activePlyr) {
-    activePlyr = new Plyr(mainVideo, {
-      autoplay: false,
-      muted: true,
-      hideControls: true,
-      controls: ['play-large', 'pip'],
-    });
-  }
-  activePlyr.source = {
-    type: 'video',
-    sources: videoInfo.source,
-  };
-  mainVideoContainer.style.display = 'block';
-  mainImg.style.display = 'none';
-}
-
-function updateImageContent(mainImg, mainVideoContainer, src) {
-  if (activePlyr) {
-    activePlyr.pause();
-  }
-  mainImg.src = src;
-  mainImg.style.display = 'block';
-  mainVideoContainer.style.display = 'none';
-}
-
-function formatNumber(number) {
-  return number.toString().padStart(2, '0');
 }
