+++ src/public/sass/ITCSS/06_components/pages/product/_details.scss
@@ -80,7 +80,6 @@
     }
 
     &__btn {
-      padding: 0.6rem 1rem;
       font-size: 1.2rem;
     }
   }
