+++ src/public/sass/ITCSS/06_components/pages/product/carousel/_carousel.scss
@@ -29,7 +29,7 @@
     display: none;
     width: 100%;
     margin-bottom: $spacing-md;
-    aspect-ratio: 5 / 3;
+    aspect-ratio: 8 / 4;
     @include desktop() {
       display: block;
     }
